Move Function to lib.rs, delete function folder
