+++ src/serialize.rs
@@ -102,28 +102,27 @@ impl EGraph {
             .filter(|(_, function)| !function.decl.ignore_viz)
             .map(|(name, function)| {
                 let mut tuples = vec![];
-                self.backend
-                    .dump_table(function.new_backend_id, |func_row| {
-                        if tuples.len() >= config.max_calls_per_function.unwrap_or(usize::MAX) {
-                            return;
-                        }
-                        let (out, inps) = func_row.vals.split_last().unwrap();
-                        use numeric_id::NumericId;
-                        tuples.push((
-                            function,
-                            inps.to_vec(),
-                            *out,
-                            func_row.subsumed,
-                            self.value_to_class_id(&function.schema.output, out),
-                            self.to_node_id(
-                                None,
-                                SerializedNode::Function {
-                                    name: *name,
-                                    offset: out.rep() as usize,
-                                },
-                            ),
-                        ));
-                    });
+                self.backend.dump_table(function.backend_id, |func_row| {
+                    if tuples.len() >= config.max_calls_per_function.unwrap_or(usize::MAX) {
+                        return;
+                    }
+                    let (out, inps) = func_row.vals.split_last().unwrap();
+                    use numeric_id::NumericId;
+                    tuples.push((
+                        function,
+                        inps.to_vec(),
+                        *out,
+                        func_row.subsumed,
+                        self.value_to_class_id(&function.schema.output, out),
+                        self.to_node_id(
+                            None,
+                            SerializedNode::Function {
+                                name: *name,
+                                offset: out.rep() as usize,
+                            },
+                        ),
+                    ));
+                });
                 tuples
             })
             // Filter out functions with no calls
