+++ tests/integration_test.rs
@@ -418,7 +418,7 @@ fn get_function(egraph: &EGraph, name: &str) -> Function {
 }
 fn get_value(egraph: &EGraph, name: &str) -> Value {
     let mut out = None;
-    let id = get_function(egraph, name).new_backend_id;
+    let id = get_function(egraph, name).backend_id;
     egraph.backend.dump_table(id, |row| out = Some(row.vals[0]));
     out.unwrap()
 }
